테이블 변경으로 인한 수정(변경내역은 설명 참조)

+++ src/main/java/com/festa/aop/CheckLoginStatusAop.java
@@ -70,7 +70,7 @@ public class CheckLoginStatusAop {
     public ResponseEntity<HttpStatus> hostLoginStatus() {
         allUserLoginStatus();
 
-        int userNo = loginService.getUserNo();
+        long userNo = loginService.getUserNo();
         MemberDTO memberInfo = memberService.getUser(userNo);
 
         log.debug(userNo + ": Started to check Host-user authentication");
